* commit 'd32da790e5bf837a9b3324e1552a11e4114ca744':
  SW-10678 - Guard against empty $DIR in clear_cache.sh

+++ cache/clear_cache.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 DIR="$(cd "$(dirname "$0")" && pwd)"
 
+## Guard against empty $DIR
+if [[ "$DIR" != */cache ]]; then
+    echo "Could not detect working directory."
+    exit 1
+fi
+
 echo "Clearing caches"
 find $DIR -mindepth 1 -maxdepth 1 -type d -print0 | xargs -0 rm -Rf
 rm -f $DIR/../web/cache/*.js > /dev/null
